Include a pre-compiled version of LLDB.framework.
The plugin uses the provided LLDB.framework, but with the system
debugserver (XCode and/or the command line tools have to be installed).
With this patch we provide one modified version of LLDB.frameworks, with
our patches.
The plugin is also changed to allow users to change a variable and use the
provided debugserver (it may have some enhancements not in the system
version).
In the future, a setting may be available to change between the bundled
version and the system version.